BackEnd/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
mysql-connector-python
SQLAlchemy
python-dotenv
torch
torchvision
ultralytics
keras
opencv-python
pillow
matplotlib
prophet
scikit-learn
numpy
pandas
